test for larger packet sizes, and larger text blocks... not successfull yet...

+++ displaySerialBuffer.cpp
@@ -15,12 +15,12 @@ std::vector<uint8_t>& displaySerialBuffer::getBuffer()
 	return m_buffer;
 }
 
-void displaySerialBuffer::setPacketIndex(uint8_t newIndex)
+void displaySerialBuffer::setPacketIndex(uint16_t newIndex)
 {
 	m_dataIndex = newIndex;
 }
 
-uint8_t displaySerialBuffer::getPacketLength() const
+uint16_t displaySerialBuffer::getPacketLength() const
 {
 	if(m_buffer.empty())
 	{
@@ -30,7 +30,7 @@ uint8_t displaySerialBuffer::getPacketLength() const
 	return *m_buffer.cbegin();
 }
 
-uint8_t displaySerialBuffer::getBufferSize() const
+uint16_t displaySerialBuffer::getBufferSize() const
 {
 	return m_buffer.size();
 }
