+++ packetProperty.hpp
@@ -60,7 +60,7 @@ class packProperty
 	//serialize an array
 	template<typename Q = T>
 	typename std::enable_if<std::is_pointer<Q>::value, bool>::type
-	serialize(displaySerialBuffer& buffer, uint8_t num_elements) const
+	serialize(displaySerialBuffer& buffer, uint16_t num_elements) const
 	{
 		if(m_flags.getFlag(m_flagId))
 		{
